requirements.txt
alembic~=1.8.0
anyio~=3.4.0
asgiref~=3.5.0
cffi~=1.15.0
cryptography~=36.0.2
fastapi~=0.92.0
greenlet~=1.1.2
gunicorn~=20.1.0
h11~=0.13.0
idna~=3.3
Jinja2~=3.1.2
Mako~=1.2.4
MarkupSafe~=2.1.1
python-multipart~=0.0.5
psycopg2~=2.9.3
pycparser~=2.21
pydantic~=1.9.0
PyJWT~=2.6.0
python-dotenv~=0.19.2
pytz~=2022.7.1
sniffio~=1.2.0
SQLAlchemy~=2.0.4
starlette~=0.25.0
typing-extensions~=4.2.0
uvicorn~=0.20.0

pip~=23.0
wheel~=0.37.1
platformdirs~=3.0.0
tomli~=2.0.1
pathspec~=0.11.0
setuptools~=58.1.0
pytz~=2022.7.1
